requirements.txt
pynetbox>=6.0.0
PyYAML>=5.4.0
click>=8.0.0
